refactor: LoadingIndicator ui.component로 이동
